Added a quicksort implementation
Tried to do a sortBy but I can't work out how to use Ordering and the
documentation assumes that I already know everything about it and all I
need is the types.

Quite a difficult language to get into, every guide and piece of
documentation needs it's own guide or documentation to even read.

Clojure plz